+++ crates/hyperdrive-math/src/lp.rs
@@ -1,3 +1,4 @@
 mod add;
+mod helpers;
 mod math;
 mod remove;
